- По истечении 2 часов пользователь не может удалять или редактировать пост

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 use App\Models\Posts\Post;
 use App\Services\Post\PostService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -43,22 +44,37 @@ class PostsController extends Controller
         $postQuery = Post::where('id', $request->id);
         $post = $postQuery->first();
 
-        if ($user->id === $post->user_id) {
+        $diffTime = $this->diffTime($post);
+
+        if ($user->id === $post->user_id && $diffTime <= 2) {
             $postQuery->delete();
             return $this->responseAPI('Комментарий удален', false);
+        }else{
+            return $this->responseAPI('Вы не можете удалить комментарий по истечении 2 часов', true);
         }
-        return $this->responseAPI('Вы не можете удалить комментарий', true);
+
     }
+
     public function editPost(Request $request){
         $user = Auth::user();
         $postQuery = Post::where('id', $request->id);
         $post = $postQuery->first();
 
-        if ($user->id === $post->user_id) {
+        $diffTime = $this->diffTime($post);
+
+        if ($user->id === $post->user_id && $diffTime <= 2) {
             $post->update(['message' => $request->message]);
             return $this->responseAPI('Комментарий редактирован', false);
+        }else{
+            return $this->responseAPI('Вы не можете редактировать комментарий по истечении 2 часов', true);
         }
-        return $this->responseAPI('Вы не можете редактировать комментарий', true);
+
+    }
+
+    public function diffTime($post){
+        $now = Carbon::now();
+        $timeCreationPost = $post->created_at;
+        return Carbon::parse($timeCreationPost)->floatDiffInHours($now);
     }
 
 }
